Extract tagged mobile attach coverage
