+++ cmuxTests/MobileHostAuthorizationTests.swift
@@ -10,109 +10,6 @@ import Testing
 @Suite(.serialized)
 @MainActor
 struct MobileHostAuthorizationTests {
-    private func loopbackRoute() throws -> CmxAttachRoute {
-        try CmxAttachRoute(
-            id: "debug_loopback",
-            kind: .debugLoopback,
-            endpoint: .hostPort(host: "127.0.0.1", port: 58465),
-            priority: 0
-        )
-    }
-
-    private func tailscaleRoute() throws -> CmxAttachRoute {
-        try CmxAttachRoute(
-            id: "tailscale",
-            kind: .tailscale,
-            endpoint: .hostPort(host: "100.64.0.5", port: 58465),
-            priority: 10
-        )
-    }
-
-    private func compactTicket(from attachURL: String) throws -> CmxAttachTicket {
-        let components = try #require(URLComponents(string: attachURL))
-        var encoded = try #require(
-            components.queryItems?.first(where: { $0.name == "payload" })?.value
-        )
-        encoded = encoded.replacingOccurrences(of: "-", with: "+")
-            .replacingOccurrences(of: "_", with: "/")
-        encoded += String(repeating: "=", count: (4 - encoded.count % 4) % 4)
-        let data = try #require(Data(base64Encoded: encoded))
-        return try CmxAttachTicketCompactCoder().decode(data)
-    }
-
-    @Test func attachTargetsSelectOnlyRoutesValidForTheirDestination() throws {
-        let loopback = try loopbackRoute()
-        let tailscale = try tailscaleRoute()
-        let routes = [loopback, tailscale]
-
-        #expect(try MobileAttachTarget.simulatorInjection.selectRoutes(from: routes) == [loopback])
-        #expect(try MobileAttachTarget.physicalDevice.selectRoutes(from: routes) == [tailscale])
-        #expect(try MobileAttachTarget.ticketOnly.selectRoutes(from: routes) == routes)
-    }
-
-    @Test func simulatorInjectionPayloadIsLosslessV1WithLoopbackToken() throws {
-        let store = MobileAttachTicketStore()
-        let ticket = try store.createTicket(
-            workspaceID: "",
-            terminalID: nil,
-            routes: [try loopbackRoute()],
-            ttl: 3600
-        )
-
-        let payload = try store.payload(for: ticket, target: .simulatorInjection)
-        let attachURL = try #require(payload["attach_url"] as? String)
-        #expect(attachURL.contains("?v=1&payload="))
-        let decoded = try compactTicket(from: attachURL)
-        #expect(decoded.routes == ticket.routes)
-        #expect(decoded.authToken == ticket.authToken)
-    }
-
-    @Test func physicalDevicePayloadIsV2WithExactTailscaleRoutes() throws {
-        let store = MobileAttachTicketStore()
-        let ticket = try store.createTicket(
-            workspaceID: "",
-            terminalID: nil,
-            routes: [try tailscaleRoute()],
-            ttl: 3600
-        )
-
-        let payload = try store.payload(for: ticket, target: .physicalDevice)
-        let attachURL = try #require(payload["attach_url"] as? String)
-        let components = try #require(URLComponents(string: attachURL))
-        #expect(components.queryItems?.first(where: { $0.name == "v" })?.value == "2")
-        #expect(components.queryItems?.contains(where: { $0.name == "payload" }) == false)
-        #expect(try CmxPairingQRCode().decode(components).routes == ticket.routes)
-    }
-
-    @Test func ticketOnlyPayloadPreservesMixedRoutesWithoutAttachURL() throws {
-        let store = MobileAttachTicketStore()
-        let routes = [try loopbackRoute(), try tailscaleRoute()]
-        let ticket = try store.createTicket(
-            workspaceID: "",
-            terminalID: nil,
-            routes: routes,
-            ttl: 3600
-        )
-
-        let payload = try store.payload(for: ticket, target: .ticketOnly)
-        #expect(payload["attach_url"] == nil)
-        #expect((payload["routes"] as? [[String: Any]])?.count == routes.count)
-    }
-
-    @Test func physicalDevicePayloadNeverFallsBackToLoopbackV1() throws {
-        let store = MobileAttachTicketStore()
-        let ticket = try store.createTicket(
-            workspaceID: "",
-            terminalID: nil,
-            routes: [try loopbackRoute()],
-            ttl: 3600
-        )
-
-        #expect(throws: MobileAttachTicketStoreError.invalidAttachURL) {
-            try store.payload(for: ticket, target: .physicalDevice)
-        }
-    }
-
     @Test func testAttachTicketStoreKeepsMultipleTicketsForSameTerminal() throws {
         let store = MobileAttachTicketStore()
         let route = try CmxAttachRoute(
